avances sobre validaciones y respuesta de api

+++ src/main/java/com/sgav/sgav/usuario/UsuarioService.java
@@ -1,6 +1,7 @@
 package com.sgav.sgav.usuario;
 
 import com.sgav.sgav.util.Helper;
+import com.sgav.sgav.util.ResponseCustom;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,41 +19,55 @@ public class UsuarioService {
     @Autowired
     private UsuarioDao usuarioDao;
 
-    public ResponseEntity<String> addUsuario(Usuario usuario){
+    ResponseCustom responseCustom = new ResponseCustom();
+
+    public ResponseEntity<?> addUsuario(Usuario usuario){
 
         if(usuario.getNombre() == null || usuario.getNombre().isEmpty()){
-            return new ResponseEntity<>("Nombre vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("Nombre vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }else{
             if(!Helper.isValidName(usuario.getNombre())){
-                return new ResponseEntity<>("Nombre no valido, solo letras sin espacios ni numeros, verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("Nombre no valido, solo letras sin espacios ni numeros, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(usuario.getApellido() == null || usuario.getApellido().isEmpty()){
-            return new ResponseEntity<>("apellido vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("apellido vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }else{
             if(!Helper.isValidName(usuario.getApellido())){
-                return new ResponseEntity<>("Apellido no valido, solo letras sin espacios ni numeros, verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("Apellido no valido, solo letras sin espacios ni numeros, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(usuario.getDni() == null || usuario.getDni() == 0){
-            return new ResponseEntity<>("dni vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("dni vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
+        }else if(!Helper.isValidDNI(usuario.getDni())){
+            responseCustom.setResponse("dni igual a 0 o negativo, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }
         if(usuario.getTelefono() == null || usuario.getTelefono().isEmpty()){
-            return new ResponseEntity<>("telefono vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("telefono vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }else{
             if(!Helper.isValidPhoneNumber(usuario.getTelefono())){
-                return new ResponseEntity<>("telefono invalido, 13 numeros maximo sin espacios ni caracteres especiales" +
-                        ", verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("telefono invalido, 13 numeros maximo sin espacios ni caracteres especiales, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(usuario.getRolId() == null || usuario.getRolId() == 0){
-            return new ResponseEntity<>("rol vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("rol vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }
         if(usuario.getSexo() == null || usuario.getSexo().isEmpty()){
-            return new ResponseEntity<>("sexo vacio, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("sexo vacio, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }
         if(usuario.getEnabled() == null){
-            return new ResponseEntity<>("habilitación vacia, verificar", HttpStatus.BAD_REQUEST);
+            responseCustom.setResponse("habilitación vacia, verificar");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
         }
 
         usuarioRepository.save(usuario);
@@ -60,36 +75,38 @@ public class UsuarioService {
         return new ResponseEntity<>("Usuario agregado", HttpStatus.OK);
     }
 
-    public ResponseEntity<String> getUsuario(Usuario usuario){
+    public ResponseEntity<?> getUsuario(Usuario usuario){
         Usuario u = usuarioDao.getUsuario(usuario.getNombre(),usuario.getApellido());
         if(u == null){
             return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>("Found", HttpStatus.OK);
+        return new ResponseEntity<>(u, HttpStatus.OK);
     }
 
-    public ResponseEntity<String> updateUsuario(Usuario usuario){
+    public ResponseEntity<?> updateUsuario(Usuario usuario){
 
         if(!Helper.isNullOrEmpty(usuario.getNombre())){
             if(!Helper.isValidName(usuario.getNombre())){
-                return new ResponseEntity<>("Nombre no valido, solo letras " +
-                        "sin espacios ni numeros, verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("Nombre no valido, solo letras sin espacios ni numeros, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(!Helper.isNullOrEmpty(usuario.getApellido())){
             if(!Helper.isValidName(usuario.getApellido())){
-                return new ResponseEntity<>("Apellido no valido, solo letras sin espacios ni numeros, verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("Apellido no valido, solo letras sin espacios ni numeros, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(usuario.getDni() != null){
-            if(usuario.getDni() <= 0){
-                return new ResponseEntity<>("dni igual a 0 o negativo, verificar", HttpStatus.BAD_REQUEST);
+            if(!Helper.isValidDNI(usuario.getDni())){
+                responseCustom.setResponse("DNI invalido, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
         if(!Helper.isNullOrEmpty(usuario.getTelefono())){
             if(!Helper.isValidPhoneNumber(usuario.getTelefono())){
-                return new ResponseEntity<>("telefono invalido, 13 numeros maximo sin espacios ni caracteres especiales" +
-                        ", verificar", HttpStatus.BAD_REQUEST);
+                responseCustom.setResponse("telefono invalido, 13 numeros maximo sin espacios ni caracteres especiales, verificar");
+                return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
             }
         }
 
@@ -97,8 +114,12 @@ public class UsuarioService {
         return new ResponseEntity<>("usuario actualizado", HttpStatus.OK);
     }
 
-    public ResponseEntity<String> deleteUsuario(Usuario usuario){
+    public ResponseEntity<?> deleteUsuario(Usuario usuario){
 
+        if(usuario.getId() == null || usuario.getId() <0){
+            responseCustom.setResponse("se requiere id para esta operación");
+            return new ResponseEntity<>(responseCustom, HttpStatus.BAD_REQUEST);
+        }
         usuarioRepository.delete(usuario);
         return new ResponseEntity<>("usuario eliminado", HttpStatus.OK);
     }
@@ -115,4 +136,5 @@ public class UsuarioService {
 
         return new ResponseEntity<>(usuarioList, HttpStatus.OK);
     }
+
 }
